Generate meta.json file in target directory. Web server only shows contents of target directory.
